requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pydantic==2.8.2
pandera==0.20.3
SQLAlchemy==2.0.36
fastapi==0.114.1
uvicorn==0.30.6
pyyaml==6.0.2
streamlit==1.38.0
python-dateutil==2.9.0.post0
loguru==0.7.2
requests==2.32.3
